+++ boards/clue_nrf52840/src/main.rs
@@ -214,7 +214,7 @@ pub unsafe fn main() {
 
     let gpio = components::gpio::GpioComponent::new(
         board_kernel,
-        capsules::gpio::DRIVER_NUM as u32,
+        capsules::gpio::DRIVER_NUM,
         components::gpio_component_helper!(
             nrf52840::gpio::GPIOPin,
             2 => &nrf52840_peripherals.gpio_port[GPIO_D2],
@@ -248,7 +248,7 @@ pub unsafe fn main() {
     //--------------------------------------------------------------------------
     let button = components::button::ButtonComponent::new(
         board_kernel,
-        capsules::button::DRIVER_NUM as u32,
+        capsules::button::DRIVER_NUM,
         components::button_component_helper!(
             nrf52840::gpio::GPIOPin,
             (
@@ -288,7 +288,7 @@ pub unsafe fn main() {
         .finalize(components::alarm_mux_component_helper!(nrf52::rtc::Rtc));
     let alarm = components::alarm::AlarmDriverComponent::new(
         board_kernel,
-        capsules::alarm::DRIVER_NUM as u32,
+        capsules::alarm::DRIVER_NUM,
         mux_alarm,
     )
     .finalize(components::alarm_component_helper!(nrf52::rtc::Rtc));
@@ -324,7 +324,7 @@ pub unsafe fn main() {
             virtual_alarm_buzzer,
             capsules::buzzer_driver::DEFAULT_MAX_BUZZ_TIME_MS,
             board_kernel.create_grant(
-                capsules::buzzer_driver::DRIVER_NUM as u32,
+                capsules::buzzer_driver::DRIVER_NUM,
                 &memory_allocation_capability
             )
         )
@@ -375,7 +375,7 @@ pub unsafe fn main() {
     // Setup the console.
     let console = components::console::ConsoleComponent::new(
         board_kernel,
-        capsules::console::DRIVER_NUM as u32,
+        capsules::console::DRIVER_NUM,
         uart_mux,
     )
     .finalize(());
@@ -388,7 +388,7 @@ pub unsafe fn main() {
 
     let rng = components::rng::RngComponent::new(
         board_kernel,
-        capsules::rng::DRIVER_NUM as u32,
+        capsules::rng::DRIVER_NUM,
         &base_peripherals.trng,
     )
     .finalize(());
@@ -429,7 +429,7 @@ pub unsafe fn main() {
         capsules::proximity::ProximitySensor<'static>,
         capsules::proximity::ProximitySensor::new(
             apds9960,
-            board_kernel.create_grant(capsules::proximity::DRIVER_NUM as u32, &grant_cap)
+            board_kernel.create_grant(capsules::proximity::DRIVER_NUM, &grant_cap)
         )
     );
 
@@ -441,14 +441,14 @@ pub unsafe fn main() {
 
     let temperature = components::temperature::TemperatureComponent::new(
         board_kernel,
-        capsules::temperature::DRIVER_NUM as u32,
+        capsules::temperature::DRIVER_NUM,
         sht3x,
     )
     .finalize(());
 
     let humidity = components::humidity::HumidityComponent::new(
         board_kernel,
-        capsules::humidity::DRIVER_NUM as u32,
+        capsules::humidity::DRIVER_NUM,
         sht3x,
     )
     .finalize(());
@@ -503,7 +503,7 @@ pub unsafe fn main() {
 
     let screen = components::screen::ScreenComponent::new(
         board_kernel,
-        capsules::screen::DRIVER_NUM as u32,
+        capsules::screen::DRIVER_NUM,
         tft,
         Some(tft),
     )
@@ -515,7 +515,7 @@ pub unsafe fn main() {
 
     let ble_radio = nrf52_components::BLEComponent::new(
         board_kernel,
-        capsules::ble_advertising_driver::DRIVER_NUM as u32,
+        capsules::ble_advertising_driver::DRIVER_NUM,
         &base_peripherals.ble_radio,
         mux_alarm,
     )
@@ -538,7 +538,7 @@ pub unsafe fn main() {
 
     let (ieee802154_radio, _mux_mac) = components::ieee802154::Ieee802154Component::new(
         board_kernel,
-        capsules::ieee802154::DRIVER_NUM as u32,
+        capsules::ieee802154::DRIVER_NUM,
         &base_peripherals.ieee802154_radio,
         aes_mux,
         PAN_ID,
@@ -572,7 +572,7 @@ pub unsafe fn main() {
         alarm: alarm,
         ipc: kernel::ipc::IPC::new(
             board_kernel,
-            kernel::ipc::DRIVER_NUM as u32,
+            kernel::ipc::DRIVER_NUM,
             &memory_allocation_capability,
         ),
         temperature: temperature,
