+++ boards/components/src/ninedof.rs
@@ -40,11 +40,11 @@ macro_rules! ninedof_component_helper {
 
 pub struct NineDofComponent {
     board_kernel: &'static kernel::Kernel,
-    driver_num: u32,
+    driver_num: usize,
 }
 
 impl NineDofComponent {
-    pub fn new(board_kernel: &'static kernel::Kernel, driver_num: u32) -> NineDofComponent {
+    pub fn new(board_kernel: &'static kernel::Kernel, driver_num: usize) -> NineDofComponent {
         NineDofComponent {
             board_kernel: board_kernel,
             driver_num: driver_num,
